feat: external modules for main, sidebar and cart container
